Add download util function
