+++ pkg/utils/package.go
@@ -1,0 +1,2 @@
+// Package utils contains some general purpose util functions.
+package utils
